refactor: change error classes exports to indexes files
