+++ src/errors/empty-properties/index.ts
@@ -1,0 +1,2 @@
+export * from './InvalidEmptyProperty'
+export * from './InvalidEmptyID'
